Remove need for avx2-32bit-common.h
